\name{DCSmooth-package}
\alias{DCSmooth-package}
\alias{DCSmooth}
\docType{package}
\title{
\packageTitle{DCSmooth}
}
\description{
\packageDescription{DCSmooth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DCSmooth}
\packageIndices{DCSmooth}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DCSmooth}

Maintainer: \packageMaintainer{DCSmooth}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
