% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Functions.R
\name{DCSmooth}
\alias{DCSmooth}
\title{Nonparametric Double Conditional Smoothing for 2D Surfaces}
\usage{
DCSmooth(Y, X = 1, T = 1, bndw = "auto", dcsOptions = setOptions())
}
\arguments{
\item{Y}{A numeric matrix that contains the observations of the random field
or functional time-series.}

\item{X}{An optional numeric vector containing the exogenous covariates
with respect to the rows.}

\item{T}{An optional numeric vector containing the exogenous covariates
with respect to the columns.}

\item{bndw}{Bandwidth for smoothing the observations in \code{Y}. Can be a
two-valued numerical vector with bandwidths in row- and column-direction.
If the value is \code{"auto"} bandwidth selection will be carried out by
the iterative plug-in algorithm.}

\item{dcsOptions}{Additional options for the smoothing procedure. Is set
via the \code{setOptions} command.}
}
\value{
\code{DCSmooth} returns an object of class "dcs".
}
\description{
\code{DCSmooth} provides a double conditional nonparametric smoothing of the
expectation surface of a functional time series or a random field on a
lattice. Bandwidth selection is done via an iterative plug-in method.
}
