% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCS_Methods.R
\name{plot.dcs}
\alias{plot.dcs}
\title{Contour Plot for the Double Conditional Smoothing}
\usage{
\method{plot}{dcs}(object, color = c("#444C5C", "#78A5A3", "#E1B16A", "#CE5A57"))
}
\description{
Contour Plot for the Double Conditional Smoothing
}
