% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCS_MainFunctions.R
\name{plotDCS}
\alias{plotDCS}
\title{3D Surface Plot of "dcs"-object or numeric matrix}
\usage{
plotDCS(DCSobj, ...)
}
\arguments{
\item{DCSobj}{An object of class "dcs" or a numeric matrix that contains the
values to be plotted.}

\item{...}{optional arguments passed to the plot function. See the vignette 
\code{vignette("DCS", package = "DCSmooth")}}
}
\value{
\code{DCSmooth} returns an object of class "plotly" and "htmlwidget".
}
\description{
\code{plotDCS} uses the plotly device to plot the 3D surface of the given
"dcs-object or matrix. If a "dcs"-object is passed to the function, it can
be chosen between plots of the original data (1), smoothed surface (2) and 
residuals (3).
}
\section{}{
NA
}

\examples{
smth =  dcs(y.norm1)
plotDCS(smth)

}
