% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_DCS.R
\name{print.dcs}
\alias{print.dcs}
\title{Summarize Results from Double Conditional Smoothing}
\usage{
\method{print}{dcs}(x, ...)
}
\arguments{
\item{x}{an object of class "dcs", usually, a result of a call to 
\code{\link{dcs}}.}

\item{...}{Additional arguments passed to \code{print.dcs}.}
}
\value{
No return value.
}
\description{
\code{print} method for class \code{"dcs"}
}
\section{Details}{

\code{print.dcs} prints a short summary of an object of class \code{dcs},
only including bandwidths and the estimated variance coefficient (only if
automatic bandwidth selection is used).
}

\examples{
y <- y.norm1 + matrix(rnorm(101^2), nrow = 101, ncol = 101)
dcs_object <- dcs(y)
print(dcs_object)
dcs_object

}
\seealso{
\code{\link{plot.dcs}}, \code{\link{print.dcs_options}}
}
