% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCS_Methods.R
\name{print.dcs_options}
\alias{print.dcs_options}
\title{Print Options for Double Conditional Smoothing}
\usage{
\method{print}{dcs_options}(object)
}
\arguments{
\item{object}{an object of class "dcs_options", usually, a result of a call to 
\code{\link{set_options}}.}
}
\description{
\code{print} method for class \code{"dcs_options"}
}
\section{Details}{

\code{print.dcs_options} prints the obptions used for the \code{\link{dcs}} 
function, which should be an object of class \code{"dcs_options"}.
}

\examples{
## Default options
myOpt <- set.options()
print(myOpt)

## Use Kernel regression
myOpt <- set.options(type = "KR")
print(myOpt)

}
\seealso{
\code{\link{print.dcs}}
}
