% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_DCS.R
\name{print.summary_dcs}
\alias{print.summary_dcs}
\title{Print the Summary of a DCS estimation}
\usage{
\method{print}{summary_dcs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary_dcs"}.}

\item{...}{Additional arguments passed to \code{print.summary_dcs}.}
}
\value{
No return value.
}
\description{
\code{print} method for class \code{"summary_dcs"}
}
\seealso{
\code{\link{summary.dcs}}
}
