% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_LM_SM.R
\name{print.summary_sarma}
\alias{print.summary_sarma}
\alias{print.summary_sfarima}
\title{Print the Summary of a "sarma"/"sfarima" object}
\usage{
\method{print}{summary_sarma}(x, ...)

\method{print}{summary_sfarima}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary_sarma"} or
\code{"summary_sfarima"}.}

\item{...}{Additional arguments passed to \code{print.summary_sarma}/
\code{print.summary_sfarima}.}
}
\value{
No return value.
}
\description{
\code{print} methods for class \code{"summary_sarma"}/
 \code{"summary_sfarima"}
}
\seealso{
\code{\link{summary.sarma}} \code{\link{summary.sfarima}}
}
