% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCS_Methods.R
\name{residuals.dcs}
\alias{residuals.dcs}
\title{Residuals of "dcs"-object}
\usage{
\method{residuals}{dcs}(dcs_object)
}
\arguments{
\item{dcs_object}{an object of class \code{"dcs"}, usually the result of a call to \code{\linl{dcs}}}
}
\description{
Returns the residuals of an object of class \code{"dcs"}.
}
\examples{
y = y.norm1 + matrix(rnorm(101^2), nrow = 101, ncol = 101)
dcs_object = dcs(y)
residuals(dcs_object)

}
\seealso{
\code{\link{dcs}}
}
