% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns.alv}
\alias{returns.alv}
\title{Returns of Allianz SE}
\format{
A numeric matrix with 2739 rows representing the days and 510 columns
representing the intraday time points.
}
\usage{
returns.alv
}
\description{
The (log-) returns of the shares of the German insurance company Allianz SE 
from 2004-01-02 to 2014-09-30 aggregated to per-minute observations. The data
is adjusted to matrix form for direct use with the \code{DCSmooth}-functions.
}
\keyword{datasets}
