% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns.sie}
\alias{returns.sie}
\title{Returns of Siemens AG}
\format{
A numeric matrix with 2739 rows representing the days and 510 columns
representing the intraday time points.
}
\usage{
returns.sie
}
\description{
The (log-) returns of shares of the German technology corporation Siemens AG 
from 2004-01-02 to 2014-09-30 aggregated to per-minute observations. The data
is adjusted to matrix form for direct use with the \code{DCSmooth}-functions.
}
\keyword{datasets}
