% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEST.R
\name{spec_test}
\alias{spec_test}
\title{Nonparametric Double Conditional Smoothing for 2D Surfaces}
\usage{
spec_test(Y, omega, lng)
}
\arguments{
\item{Y}{A numeric matrix that contains the observations of the random field
or functional time-series.}

\item{omega}{local point}

\item{lng}{language}
}
\description{
\code{spec_test}
}
