% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp.yuma}
\alias{temp.yuma}
\title{Temperatures from Yuma, AZ}
\format{
A numeric matrix with 365 rows and 288 columns containing the
temperatures in Celsius.
}
\usage{
temp.yuma
}
\description{
This dataset contains the 5-minute observations of the 2020 temperature in
Yuma, AZ. The data is from the U.S. Climate Reference Network database at 
\url{https://www.ncdc.noaa.gov/crn/qcdatasets.html}. (see Diamond et. al.
(2013), \href{https://journals.ametsoc.org/view/journals/bams/94/4/bams-d-12-00170.1.xml}{doi:10.1175/BAMS-D-12-00170.1}).
The observations were adjusted to only include complete days and are fitted 
to matrix form for direct use with the \code{DCSmooth}-functions.
}
\keyword{datasets}
