% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y.norm1}
\alias{y.norm1}
\title{Single Gaussian Peak}
\format{
A numeric matrix with 101 rows and 101 columns.
}
\usage{
y.norm1
}
\description{
Example data for using the DCSmooth functions. Data resembles a single
  gaussian peak on the interval \eqn{[0,1] \times [0,1]}{[0,1] x [0,1]}
  with maximum at \eqn{(0.5, 0.5)} and variance matrix \eqn{0.1 \cdot
 {\bf I}}{0.05 I}, where 
  \eqn{{\bf I}}{I} represents the \eqn{2 \times 2}{2 x 2} identity matrix.
}
\keyword{datasets}
