% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y.norm2}
\alias{y.norm2}
\title{Double Gaussian Peak}
\format{
A numeric matrix with 100 rows and 100 columns.
}
\usage{
y.norm2
}
\description{
Example data for using the DCSmooth functions. Data resembles two gaussian
  peaks on the interval \eqn{[0,1] \times [0,1]}{[0,1] x [0,1]} with maxima
  at \eqn{(0.5, 0.3)} with variance matrix \eqn{0.1 \cdot \mathbb{I}}{0.1 I}
  and at \eqn{(0.2, 0.8)} with variance matrix \eqn{0.05 \cdot \mathbb{I}}{
  0.05 I}.
}
\keyword{datasets}
