% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y.norm3}
\alias{y.norm3}
\title{Double Gaussian Ridges}
\format{
A numeric matrix with 101 rows and 101 columns.
}
\usage{
y.norm3
}
\description{
Example data for using the DCSmooth functions. Data resembles two gaussian
  ridges on the interval \eqn{[0,1] \times [0,1]}{[0,1] x [0,1]} with maxima
  at \eqn{(0.25, 0.75)} with variance matrix
  \eqn{(0.01, -0.1) \cdot {\bf I}}{(0.01, -0.1) I} and at
  \eqn{(0.75, 0.5)} with variance matrix
  \eqn{(0.01, -0.1) \cdot {\bf I}}{(0.01, -0.1) I}, where 
  \eqn{{\bf I}}{I} represents the \eqn{2 \times 2}{2 x 2} identity matrix.
}
\keyword{datasets}
