\name{DCSmooth-package}
\alias{DCSmooth-package}
\alias{DCSmooth}
\docType{package}
\title{\packageTitle{DCSmooth}}
\description{\packageDescription{DCSmooth}}
\section{Package Content}{\packageIndices{DCSmooth}}
\author{\packageAuthor{DCSmooth}}
\section{Maintainer}{\packageMaintainer{DCSmooth}}
\keyword{package}
