% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_DCS.R
\name{print.dcs_options}
\alias{print.dcs_options}
\title{Print and Summarize Options for Double Conditional Smoothing}
\usage{
\method{print}{dcs_options}(x, ...)
}
\arguments{
\item{x}{an object of class "dcs_options", usually, a result of a call
to \code{\link{set.options}}.}

\item{...}{Additional arguments passed to \code{print.dcs_options}.}
}
\value{
No return value.
}
\description{
\code{print} method for class \code{"dcs_options"}
}
\section{Details}{

\code{print.dcs_options} prints the main options and 
\code{summary.dcs_options} prints main and advanced (IPI) options used for
the \code{\link{dcs}} function. Arguments should be an object of class
\code{"dcs_options"}.
}

\examples{
## Default options
myOpt <- set.options()
print(myOpt)
summary(myOpt)

## Use Kernel regression
myOpt <- set.options(type = "KR")
print(myOpt)
summary(myOpt)

}
\seealso{
\code{\link{print.dcs}}, \code{\link{summary.dcs_options}}
}
