% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QARMA_Estimation.R
\name{qarma.sim}
\alias{qarma.sim}
\title{Simulation of a \eqn{QARMA(p, q)}-process on a lattice.}
\usage{
qarma.sim(nX, nT, model = list(ar, ma, sigma))
}
\arguments{
\item{Y}{A numeric matrix that contains the demeaned observations of the
random field or functional time-series.}

\item{model_order}{A list containing the orders of the QARMA model in the
form \code{model_order = list(ar = c(p1, p2), ma = c(q1, q2))}. Default
value is a \eqn{QARMA((1, 1), (1, 1))} model.}
}
\value{
The function returns a list including

\describe{
\item{ar}{A \eqn{(p_1 + 1) \times (p_2 + 1)}{(p1 + 1) x (p2 + 1)}-matrix
containing the estimated AR-parameters \eqn{\phi} of the QARMA-process. The
\eqn{[i, j]}th entry is the \eqn{(p_1 + 1 - i, p_2 + 1 - j)}th lag with the
\eqn{[p_1 + 1, p_2 + 1]}th entry being 1.}
\item{ma}{A \eqn{(q_1 + 1) \times (q_2 + 1)}{(q1 + 1) x (q2 + 1)}-matrix
containing the estimated MA-parameters \eqn{\theta} of the QARMA-process. The
\eqn{[i, j]}th entry is the \eqn{(q_1 + 1 - i, q_2 + 1 - j)}th lag with the 
\eqn{[q_1 + 1, q_2 + 1]}th entry being 1.}
\item{sigma}{The estimated standard deviation \eqn{\sigma} of the QARMA-model.}
\item{innov}{The matrix of innovations resulting from the QARMA estimation. 
The initial values (from \eqn{i = 1,...,p_1}{i = 1,...,p1} and
\eqn{j = 1,...,p_2}{j = 1,...,p2}) are set to zero.} 
}
}
\description{
The MA- and AR-parameters of a top-left quadrant ARMA process are estimated
using the Hannen-Rissanen Algorithm (Hannen-Rissanen ???). The lag-orders of 
the \eqn{QARMA(p, q)} are given by \eqn{p = (p_1, p_2), q = (q_1, q_2)}{p =
(p1, p2), q = (q1, q2)}, where \eqn{p_1, q_1}{p1, q1} are the lags over the
rows and \eqn{p_2, q_2}{p2, q2} are the lags over the columns. The estimation
process is based on the model
\deqn{\phi(B_{1}B_{2})X_{i,j} = \theta(B_{1}B_{2})u_{i,j}}{\phi(B1 B2)
X[i,j] = \theta(B1 B2)u[i,j]}
}
\section{Usage}{

\code{qarma(Y, model_order = list(ar = c(1, 1), ma = c(1, 1)))}
}

\section{Details}{

???
}

\examples{
qarma.example1
qarma.example2
qarma.est(qarma.example1)
qarma.est(qarma.example2, model_order = list(ar = c(1, 1), ma = c(0, 0))

}
