% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_DCS.R
\name{residuals.dcs}
\alias{residuals.dcs}
\title{Residuals of "dcs"-object}
\usage{
\method{residuals}{dcs}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"dcs"}, usually the result of a call to
\code{\link{dcs}}.}

\item{...}{Additional arguments passed to \code{residuals.dcs}.}
}
\value{
Returns the \eqn{n_x \times n_t}{n_x x n_t}-matrix of residuals.
}
\description{
Returns the residuals of an object of class \code{"dcs"}.
}
\examples{
y = y.norm1 + matrix(rnorm(101^2), nrow = 101, ncol = 101)
dcs_object = dcs(y)
residuals(dcs_object)

}
\seealso{
\code{\link{dcs}}
}
