% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns.alv}
\alias{returns.alv}
\title{#' Solar Radiation from Yuma, AZ
#' 
#' This dataset contains the 5-minute observations of the 2020 solar radiation
#' in Yuma, AZ. The data is from the U.S. Climate Reference Network database at 
#' \url{https://www.ncdc.noaa.gov/crn/qcdatasets.html}. (see Diamond et. al.
#' (2013), \href{https://journals.ametsoc.org/view/journals/bams/94/4/bams-d-12-00170.1.xml}{doi:10.1175/BAMS-D-12-00170.1}).
#' The observations were adjusted matrix form for direct use with the 
#' \code{DCSmooth}-functions.
#' 
#' @format A numeric matrix with 366 rows and 288 columns containing the
#' solar radiation in \eqn{W/m^2}.
"sun.yuma"}
\format{
A numeric matrix with 1016 rows representing the days and 101 columns
representing the intraday time points.
}
\usage{
returns.alv
}
\description{
#' Solar Radiation from Nunn, CO
#' 
#' This dataset contains the 5-minute observations of the 2020 solar radiation
#' in Nunn, CO. The data is from the U.S. Climate Reference Network database at 
#' \url{https://www.ncdc.noaa.gov/crn/qcdatasets.html}. (see Diamond et. al.
#' (2013), \href{https://journals.ametsoc.org/view/journals/bams/94/4/bams-d-12-00170.1.xml}{doi:10.1175/BAMS-D-12-00170.1}).
#' The observations were adjusted to matrix form for direct use with the 
#' \code{DCSmooth}-functions.
#' 
#' @format A numeric matrix with 366 rows and 288 columns containing the
#' solar radiation in \eqn{W/m^2}.
"sun.nunn"
Returns of Allianz SE
}
\details{
The (log-) returns of the shares of the German insurance company Allianz SE
from 2007-01-02 to 2010-12-30 aggregated to 5-minute observations. The data
is adjusted to matrix form for direct use with the \code{DCSmooth}-functions.
}
\keyword{datasets}
