% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns.alv}
\alias{returns.alv}
\title{Returns of Allianz SE}
\format{
A numeric matrix with 1016 rows representing the days and 101 columns
representing the intraday time points.
}
\usage{
returns.alv
}
\description{
The (log-) returns of the shares of the German insurance company Allianz SE
from 2007-01-02 to 2010-12-30 aggregated to 5-minute observations. The data
is adjusted to matrix form for direct use with the \code{DCSmooth}-functions.
}
\keyword{datasets}
