% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCS_MainFunctions.R
\name{setOptions}
\alias{setOptions}
\title{Set Options for the DCS procedure}
\usage{
setOptions(...)
}
\arguments{
\item{kernPar}{kernel parameters \eqn{k} and \eqn{\mu} as vector.}

\item{pOrder}{order of polynomials used for smoothing with default value 1 
(local linear regression). Kernel regression is done via \code{pOrder = 0}.
Orders are assumed to be the same in both directions.}

\item{inflExp}{inflation exponents for bandwidth selection.}

\item{inflPar}{inflation parameters for bandwidth selection.}

\item{delta}{shrink parameters for derivatives.}

\item{constWindow}{use a constant window width for estimation.}

\item{varEst}{procedure for estimation of variance factor \eqn{c_f}.}
}
\description{
Set Options for the DCS procedure
}
\examples{
myOpt = setOptions(kernPar = c(4,2), pOrder = 0, inflExp = c(0.5, 0.5),
inflPar = c(2, 1), delta = c(0.05, 0.05), constWindow = TRUE,
varEst = "iid")
}
