% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volumes.bmw}
\alias{volumes.bmw}
\title{Volumes of BMW AG}
\format{
A numeric matrix with 1016 rows representing the days and 102 columns
representing the intraday time points.
}
\usage{
volumes.bmw
}
\description{
The trading volumes of shares of the German car manufacturer BMW AG from
from 2007-01-02 to 2010-09-30 aggregated to 5-minute observations. The data
is adjusted to matrix form for direct use with the \code{DCSmooth}-functions.
}
\keyword{datasets}
