% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind.yuma}
\alias{wind.yuma}
\title{Wind Speed from Yuma, AZ}
\format{
A numeric matrix with 366 rows and 288 columns containing the
wind speeds in \eqn{m/s}.
}
\usage{
wind.yuma
}
\description{
This dataset contains the 5-minute observations of the 2020 wind speed in
Yuma, AZ. The data is from the U.S. Climate Reference Network database at 
\href{https://www.ncdc.noaa.gov/crn/qcdatasets.html}{www.ncdc.noaa.gov}. 
(see Diamond et al., 2013). The observations were adjusted matrix form for 
direct use with the \code{DCSmooth}-functions.
}
\keyword{datasets}
