% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y.norm2}
\alias{y.norm2}
\title{Double Gaussian Peak}
\format{
A numeric matrix with 101 rows and 101 columns.
}
\usage{
y.norm2
}
\description{
Example data for using the DCSmooth functions. Data resembles two gaussian
  peaks on the interval \eqn{[0,1] \times [0,1]}{[0,1] x [0,1]} with maxima
  at \eqn{(0.5, 0.3)} with variance matrix \eqn{0.1 \cdot {\bf I}}{0.1 I}
  and at \eqn{(0.2, 0.8)} with variance matrix \eqn{0.05 \cdot {\bf I}}{
  0.05 I}, where \eqn{{\bf I}}{I} represents the \eqn{2 \times 2}{2 x 2}
  identity matrix.
}
\keyword{datasets}
